Better solution to ambiguity problem; immediate versions of some methods
The ambiguity problem arose because both features and new symbols can be
represented by the token ID, so the parser could not figure out whether an
ID at the beginning of an input was meant to be one or the other. The
solution was to create a new token, features_or_new_symbols, that parses
everything as if it were both a features and a new_symbols and puts both
interpretations in a list. Then, when it comes across a grammar line that
can disambiguate, it chooses one interpretation and discards the other.

contradictsi, editi, and updatei are the same as their nonimmediate
versions, except they take dictionaries directly instead of phonemes.